Rename README.md to README Coding Challenge.md
Added Api and Test project.
Swagger is integrated to test the api exposed. The api exposed will take string and give back the result.

docker-compose file consist of two service, one for the API and other one from the Test
this will enable to execute docker-compose run test which will return the number of test cases with passed and failed count.